OAK-12010 Simplified index management
