+++ oak-core/src/test/java/org/apache/jackrabbit/oak/plugins/index/diff/DiffIndexTest.java
@@ -22,26 +22,19 @@ import static org.apache.jackrabbit.oak.plugins.index.IndexConstants.TYPE_PROPER
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.mockStatic;
 import static org.mockito.Mockito.when;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.List;
-import java.util.Optional;
 
-import org.apache.commons.io.IOUtils;
 import org.apache.jackrabbit.JcrConstants;
 import org.apache.jackrabbit.oak.api.Blob;
 import org.apache.jackrabbit.oak.api.CommitFailedException;
 import org.apache.jackrabbit.oak.api.PropertyState;
 import org.apache.jackrabbit.oak.api.Type;
-import org.apache.jackrabbit.oak.plugins.memory.BinaryPropertyState;
 import org.apache.jackrabbit.oak.commons.json.JsonObject;
 import org.apache.jackrabbit.oak.plugins.index.AsyncIndexUpdate;
 import org.apache.jackrabbit.oak.plugins.index.CompositeIndexEditorProvider;
@@ -49,68 +42,20 @@ import org.apache.jackrabbit.oak.plugins.index.IndexConstants;
 import org.apache.jackrabbit.oak.plugins.index.IndexEditorProvider;
 import org.apache.jackrabbit.oak.plugins.index.IndexUpdateProvider;
 import org.apache.jackrabbit.oak.plugins.index.counter.NodeCounterEditorProvider;
-import org.apache.jackrabbit.oak.plugins.index.optimizer.DiffIndexUpdater;
 import org.apache.jackrabbit.oak.plugins.index.property.PropertyIndexEditorProvider;
 import org.apache.jackrabbit.oak.plugins.index.reference.ReferenceEditorProvider;
+import org.apache.jackrabbit.oak.plugins.memory.BinaryPropertyState;
 import org.apache.jackrabbit.oak.plugins.memory.MemoryNodeStore;
 import org.apache.jackrabbit.oak.spi.commit.CommitInfo;
 import org.apache.jackrabbit.oak.spi.commit.EditorHook;
 import org.apache.jackrabbit.oak.spi.state.NodeBuilder;
 import org.apache.jackrabbit.oak.spi.state.NodeStore;
 import org.junit.Test;
-import org.mockito.MockedStatic;
 
 /**
  * Tests for DiffIndex functionality.
  */
 public class DiffIndexTest {
-
-    @Test
-    public void testFindMatchingIndexName() throws IOException {
-        String indexJson = "{\n" +
-            "  \"index\": {\n" +
-            "    \"compatVersion\": 2,\n" +
-            "    \"async\": \"async\",\n" +
-            "    \"queryPaths\": [\"/content/dam/test\"],\n" +
-            "    \"includedPaths\": [\"/content/dam/test\"],\n" +
-            "    \"jcr:primaryType\": \"nam:oak:QueryIndexDefinition\",\n" +
-            "    \"evaluatePathRestrictions\": true,\n" +
-            "    \"type\": \"lucene\",\n" +
-            "    \"tags\": [\"fragments\"],\n" +
-            "    \"indexRules\": {\n" +
-            "      \"jcr:primaryType\": \"nam:nt:unstructured\",\n" +
-            "      \"dam:Asset\": {\n" +
-            "        \"jcr:primaryType\": \"nam:nt:unstructured\",\n" +
-            "        \"properties\": {\n" +
-            "          \"jcr:primaryType\": \"nam:nt:unstructured\",\n" +
-            "          \"title\": {\n" +
-            "            \"name\": \"str:jcr:title\",\n" +
-            "            \"propertyIndex\": true,\n" +
-            "            \"jcr:primaryType\": \"nam:nt:unstructured\"\n" +
-            "          }\n" +
-            "        }\n" +
-            "      }\n" +
-            "    }\n" +
-            "  }\n" +
-            "}";
-
-        try (MockedStatic<?> mockedStatic = mockStatic(RootIndexesListService.class)) {
-            NodeStore store = mock(NodeStore.class);
-
-            String indexesJsonString;
-
-            try (InputStream stream = getClass().getResourceAsStream("/org/apache/jackrabbit/oak/plugins/index/diff/indexes.json")) {
-                indexesJsonString = IOUtils.toString(stream, StandardCharsets.UTF_8);
-            }
-
-            mockedStatic.when(() -> RootIndexesListService.getRootIndexDefinitions(eq(store), anyString()))
-                .thenReturn(JsonObject.fromJson(indexesJsonString, true));
-
-            Optional<String> matchingIndexName = DiffIndexUpdater.findMatchingIndexName(store, indexJson);
-
-            assertTrue(matchingIndexName.isPresent());
-        }
-    }
 
     @Test
     public void listIndexes() {
